refactor: move createConversation action to HomeController
