+++ app/Http/Controllers/ConversationController.php
@@ -45,60 +45,4 @@ final readonly class ConversationController
             'persona' => $firstMsg->meta['persona'] ?? '',
         ]);
     }
-
-    public function submit(HomeIndexRequest $request, SatscribeAction $action): JsonResponse
-    {
-        $search = $this->getPromptInput($request);
-        $persona = $this->getPromptPersona($request);
-        $question = $request->getQuestionInput();
-        $refreshEnabled = $request->isRefreshEnabled();
-
-        try {
-            $generatedPrompt = $action->execute($search, $persona, $refreshEnabled, $question);
-        } catch (BlockchainException|OpenAIError $e) {
-            Log::error('Failed to describe prompt result', [
-                'search' => $search->text,
-                'refreshEnabled' => $refreshEnabled,
-                'question' => $question,
-                'persona' => $persona->value,
-                'error' => $e->getMessage(),
-            ]);
-            return response()->json([
-                'error' => $e->getMessage()
-            ], Response::HTTP_UNPROCESSABLE_ENTITY);
-        }
-
-        return response()->json([
-            'maxBitcoinBlockHeight' => $this->heightProvider->getMaxPossibleBlockHeight(),
-            'search' => $search,
-            'conversationUlid' => $generatedPrompt->conversation->ulid,
-            'html' => view('partials.conversation', [
-                'conversation' => $generatedPrompt->conversation,
-                'suggestions' => $this->generateSuggestions($search),
-            ])->render()
-        ]);
-    }
-
-    private function getPromptInput(HomeIndexRequest $request): PromptInput
-    {
-        if ($request->hasSearchInput()) {
-            return PromptInput::fromRaw($request->getSearchInput());
-        }
-
-        return PromptInput::fromRaw($this->heightProvider->getCurrentBlockHeight());
-    }
-
-    private function getPromptPersona(HomeIndexRequest $request): PromptPersona
-    {
-        return PromptPersona::tryFrom($request->getPersonaInput())
-            ?? PromptPersona::from(PromptPersona::DEFAULT);
-    }
-
-    private function generateSuggestions(PromptInput $search): array
-    {
-        if ($search->isBlock()) {
-            return QuestionPlaceholder::forBlock();
-        }
-        return QuestionPlaceholder::forTx();
-    }
 }
